chore: address comments regards tests
