+++ lib/grandpa/vote_message.go
@@ -42,7 +42,7 @@ func (s *Service) sendTelemetryVoteMessage(vm *VoteMessage) {
 			),
 		)
 	default:
-		logger.Warnf("unsupported stage %s", vm.Message.Stage.String())
+		logger.Warnf("unsupported stage %s", vm.Message.Stage)
 	}
 }
 
@@ -62,11 +62,11 @@ func (s *Service) createSignedVoteAndVoteMessage(vote *Vote, stage Subround) (*S
 		return nil, nil, err
 	}
 
-	publicBytes := s.keypair.Public().(*ed25519.PublicKey).AsBytes()
+	publicKeyBytes := s.keypair.Public().(*ed25519.PublicKey).AsBytes()
 	pc := &SignedVote{
 		Vote:        *vote,
 		Signature:   ed25519.NewSignatureBytes(sig),
-		AuthorityID: publicBytes,
+		AuthorityID: publicKeyBytes,
 	}
 
 	sm := &SignedMessage{
@@ -74,7 +74,7 @@ func (s *Service) createSignedVoteAndVoteMessage(vote *Vote, stage Subround) (*S
 		BlockHash:   pc.Vote.Hash,
 		Number:      pc.Vote.Number,
 		Signature:   ed25519.NewSignatureBytes(sig),
-		AuthorityID: publicBytes,
+		AuthorityID: publicKeyBytes,
 	}
 
 	vm := &VoteMessage{
